Refactor: Update imports and replace deprecated utility functions in BulletPlanReasoner
