+++ main.py
@@ -29,7 +29,7 @@ SETUP INSTRUCTIONS:
 
 import argparse
 import logging
-import os, json
+import os
 import sys
 
 from dotenv import load_dotenv
@@ -42,14 +42,11 @@ from jentic_agents.agents.simple_ui_agent import SimpleUIAgent
 from jentic_agents.memory.scratch_pad import ScratchPadMemory
 from jentic_agents.communication.inbox.cli_inbox import CLIInbox
 from jentic_agents.communication.hitl.cli_intervention_hub import CLIInterventionHub
-from jentic_agents.memory.agent_memory import create_agent_memory
 from jentic_agents.platform.jentic_client import JenticClient
 from jentic_agents.reasoners.bullet_list_reasoner import BulletPlanReasoner
-from jentic_agents.reasoners.freeform_reasoner import FreeformReasoner
-from jentic_agents.reasoners.standard_reasoner import StandardReasoner
 
 from jentic_agents.utils.llm import LiteLLMChatLLM
-from jentic_agents.utils.config import get_config, get_config_value
+from jentic_agents.utils.config import get_config_value
 
 logging.getLogger("litellm").setLevel(logging.WARNING)
 logging.getLogger("LiteLLM").setLevel(logging.WARNING)
